Create January_Data
This is the data from the Month of January 2012.  There are 2 main types of errors in the data: the calls are not always in order by date, or the calls have the incorrect date. It needs to be split into these fields: "Address", "City", "State", "Type", "Location", "Date", "Day of Week", "Time Received", "Time Dispatched", "Time Arrived", "Time Cleared".